remove ema
