+++ orb_models/forcefield/calculator.py
@@ -1,10 +1,9 @@
-from ase.calculators.calculator import Calculator, all_changes
 from typing import Optional
+
 import torch
-from orb_models.forcefield.atomic_system import (
-    SystemConfig,
-    ase_atoms_to_atom_graphs,
-)
+from ase.calculators.calculator import Calculator, all_changes
+
+from orb_models.forcefield.atomic_system import SystemConfig, ase_atoms_to_atom_graphs
 from orb_models.forcefield.graph_regressor import GraphRegressor
