+++ orb_models/forcefield/reference_energies.py
@@ -1,4 +1,5 @@
 from typing import NamedTuple
+
 import numpy
